+++ Source/Krypton Components/Krypton.Toolkit/General/ThemeChangeCoordinator.cs
@@ -147,6 +147,12 @@ internal static class ThemeChangeCoordinator
     {
         public bool PreFilterMessage(ref Message m)
         {
+            // Don't interfere with designer operations
+            if (IsAnyFormInDesignMode())
+            {
+                return false;
+            }
+
             if (!InProgress)
             {
                 return false;
@@ -185,6 +191,31 @@ internal static class ThemeChangeCoordinator
                 return true; // swallow
             }
 
+            return false;
+        }
+
+        /// <summary>
+        /// Checks if any form in the application is currently in design mode.
+        /// </summary>
+        /// <returns>True if any form is in design mode; otherwise false.</returns>
+        private static bool IsAnyFormInDesignMode()
+        {
+            try
+            {
+                // Check all open forms to see if any are in design mode
+                foreach (Form form in Application.OpenForms)
+                {
+                    if (form.Site?.DesignMode == true)
+                    {
+                        return true;
+                    }
+                }
+            }
+            catch
+            {
+                // If we can't check, assume not in design mode to avoid breaking functionality
+            }
+            
             return false;
         }
     }